rename celerytasks -> tasks
