+++ complete.cfg
@@ -22,7 +22,7 @@ file = ${buildout:directory}/var/supervisord.sock
 programs = 
     10 zeo ${zeo:location}/bin/runzeo ${zeo:location}
     20 solr /usr/bin/env [java -Xms512m -Xmx2048m -jar start.jar] ${solr:location} true
-    30 celery ${buildout:directory}/bin/celery [-A ploneintranet.async.celerytasks worker] ${buildout:directory}/var/ true  
+    30 celery ${buildout:directory}/bin/celery [-A ploneintranet.async.tasks worker] ${buildout:directory}/var/ true  
     100 instance ${buildout:directory}/bin/instance [console] ${instance:location} true
     110 instance2 ${buildout:directory}/bin/instance2 [console] ${instance2:location} true
